BAUF-106 Zaslon za My Jobs na frontendu pretvoren u Pending Jobs
Zašto: https://github.com/AIR-DavidoviTigrovi/baufind/pull/39
Što: https://vlovric21.atlassian.net/wiki/spaces/baufind/pages/43941890/07.01.2025.
