+++ Software/Frontend/app/src/main/java/hr/foi/air/baufind/MainActivity.kt
@@ -19,12 +19,7 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import com.google.gson.Gson
-import hr.foi.air.baufind.core.map.MapProvider
-import hr.foi.air.baufind.example_map.ExampleMapProvider
-import hr.foi.air.baufind.google_map.GoogleMapProvider
-import hr.foi.air.baufind.helpers.MapHelper
 import hr.foi.air.baufind.navigation.BottomNavigationBar
-import hr.foi.air.baufind.open_street_map.OpenStreetMapProvider
 import hr.foi.air.baufind.ui.screens.JobCreateScreen.JobAddSkillsScreen
 import hr.foi.air.baufind.ui.screens.JobCreateScreen.JobDetailsScreen
 import hr.foi.air.baufind.ui.screens.JobCreateScreen.JobPositionsLocationScreen
@@ -37,6 +32,8 @@ import hr.foi.air.baufind.ui.screens.JobSearchScreen.JobSearchViewModel
 import hr.foi.air.baufind.ui.screens.LoginScreen.LoginScreen
 import hr.foi.air.baufind.ui.screens.MyJobsScreen.MyJobsScreen
 import hr.foi.air.baufind.ui.screens.MyJobsScreen.MyJobsViewModel
+import hr.foi.air.baufind.ui.screens.PendingJobsScreen.PendingJobsScreen
+import hr.foi.air.baufind.ui.screens.PendingJobsScreen.PendingJobsViewModel
 import hr.foi.air.baufind.ui.screens.Settings.SettingsScreen
 import hr.foi.air.baufind.ui.screens.UserProfileScreen.EditProfileScreen
 import hr.foi.air.baufind.ui.screens.UserProfileScreen.ReviewsScreen
@@ -57,6 +54,7 @@ class MainActivity : ComponentActivity() {
         val jobViewModel = JobViewModel()
         val jobSearchViewModel = JobSearchViewModel()
         val jobSearchDetailsViewModel = JobSearchDetailsViewModel()
+        val pendingJobsViewModel = PendingJobsViewModel()
         val myJobsViewModel = MyJobsViewModel()
         setContent {
             val navController = rememberNavController()
@@ -74,6 +72,7 @@ class MainActivity : ComponentActivity() {
                                 "jobSearchScreen",
                                 "jobSearchDetailsScreen",
                                 "settingsScreen",
+                                "pendingJobsScreen",
                                 "myJobsScreen"
                             )
                         ) {
@@ -147,6 +146,7 @@ class MainActivity : ComponentActivity() {
                             composable("jobAddSkillsScreen") { JobAddSkillsScreen(navController, jobViewModel, tokenProvider) }
 
                             composable("jobSearchScreen") { JobSearchScreen(navController, tokenProvider, jobSearchViewModel, jobSearchDetailsViewModel) }
+                            composable("pendingJobsScreen") { PendingJobsScreen(navController, tokenProvider, pendingJobsViewModel) }
                             composable("myJobsScreen") { MyJobsScreen(navController, tokenProvider, myJobsViewModel) }
                             composable("jobSearchDetailsScreen") { JobSearchDetailsScreen(navController, tokenProvider, jobSearchDetailsViewModel) }
                             composable("settingsScreen") { SettingsScreen(navController) }
